pyproject.toml
[tool.poetry]
name = "maple-infra"
version = "0.1.0"
description = "AWS Infrastructure for Maple"
authors = ["Alex Ball <me@alexjball.com>"]
packages = [{ include = 'maple' }]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.82.0"
constructs = "^10.1.268"
aws-cdk-aws-apigatewayv2-integrations-alpha = "^2.82.0a0"
aws-cdk-aws-apigatewayv2-alpha = "^2.82.0a0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
